Move checkout registration action to the integration manager
