+++ web/app/integration-manager/checkout/commands.js
@@ -12,3 +12,4 @@ export const fetchShippingMethodsEstimate = (...args) => connector.fetchShipping
 export const fetchCheckoutPaymentData = (...args) => connector.fetchCheckoutPaymentData(...args)
 export const submitPayment = (...args) => connector.submitPayment(...args)
 export const fetchCheckoutConfirmationData = (...args) => connector.fetchCheckoutConfirmationData(...args)
+export const checkoutRegister = (...args) => connector.checkoutRegister(...args)
